modal styles fix, logic refact
