+++ src/css/modal.css
@@ -1,9 +1,3 @@
-.backdrop.is-hidden,
-.rate-backdrop.is-hidden {
-  opacity: 0;
-  pointer-events: none;
-}
-
 .backdrop,
 .rate-backdrop {
   position: fixed;
@@ -15,7 +9,13 @@
   background-color: var(--bgr-backdrop);
   opacity: 1;
   overflow-y: scroll;
-  transition: opacity 250ms cubic-bezier(0.4, 0, 0.2, 1);
+  transition: opacity 0.25s cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.backdrop.is-hidden,
+.rate-backdrop.is-hidden {
+  opacity: 0;
+  pointer-events: none;
 }
 
 .modal-close-btn,
@@ -28,19 +28,21 @@
   top: 14px;
   right: 14px;
   padding: 0;
-  transition-property: transform;
-  transition-duration: 250ms;
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
 }
 .modal-close-btn svg,
 .rate-modal-close-btn svg {
   stroke: var(--white);
   transform: scale(1);
+  transition: scale 0.25s cubic-bezier(0.4, 0, 0.2, 1),
+    stroke 0.25s cubic-bezier(0.4, 0, 0.2, 1);
 }
 
 .modal-close-btn:hover svg,
-.rate-modal-close-btn:hover svg {
+.modal-close-btn:focus svg,
+.rate-modal-close-btn:hover svg,
+.rate-modal-close-btn:focus svg {
   transform: scale(1.1);
+  stroke: var(--gold);
 }
 
 .modal {
@@ -55,7 +57,20 @@
   background-color: var(--black);
   border-radius: 20px;
   position: relative;
-  transform: translate(0, 0) scale(0.9);
+}
+
+@media screen and (max-width: 767px) {
+  .modal {
+    transform: scale(1.1);
+    opacity: 0;
+    transition: transform 0.5s cubic-bezier(0.4, 0, 0.2, 1),
+      opacity 0.5s cubic-bezier(0.4, 0, 0.2, 1);
+  }
+
+  .backdrop:not(.is-hidden) .modal {
+    transform: scale(0.9);
+    opacity: 1;
+  }
 }
 
 .modal-info {
@@ -76,7 +91,7 @@
 }
 
 .exercise-title {
-  margin: 10px 0;
+  padding: 10px 0;
   color: var(--white);
   font-feature-settings: 'liga' off, 'clig' off;
   font-family: 'DM Sans';
@@ -133,9 +148,13 @@
   border-radius: 30px;
   border: 1px solid;
   background-color: var(--white);
-  transition-property: stroke fill color background-color border-color;
-  transition-duration: 250ms;
+  transition-property: stroke, fill, color, background-color, border-color;
+  transition-duration: 0.25s;
   transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+  svg {
+    fill: var(--white);
+    stroke: var(--black);
+  }
 }
 
 .modal-btn:hover {
@@ -157,11 +176,6 @@
 .modal-btn.rate-btn:hover {
   color: var(--black);
   background-color: var(--white);
-}
-
-.modal-btn svg {
-  fill: var(--white);
-  stroke: var(--black);
 }
 
 .exercise-set {
@@ -237,11 +251,13 @@
     position: absolute;
     top: 50%;
     left: 50%;
-    transform: translate(-50%, -50%) scale(1);
+    transform: translate(-50%, -50%) scale(1.1);
 
-    transition-property: transform;
-    transition-duration: 250ms;
-    transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
+    transition: transform 0.5s cubic-bezier(0.4, 0, 0.2, 1);
+  }
+
+  .backdrop:not(.is-hidden) .modal {
+    transform: translate(-50%, -50%) scale(1);
   }
 
   .modal-info {
@@ -273,7 +289,7 @@
   }
 }
 
-@media screen and (min-width: 1200px) {
+@media screen and (min-width: 1440px) {
   .modal {
     width: 708px;
   }
