+++ src/js/modal/renderExercise.js
@@ -1,12 +1,11 @@
-import axios from 'axios';
 import { refs } from '../refs';
-import storageService from '../services/storage';
 import spriteUrl from '../../images/sprite.svg';
-
-axios.defaults.baseURL = 'https://your-energy.b.goit.study/api/';
-
-const getExerciseById = (id) =>
-  axios.get(`/exercises/${id}`).then((res) => res.data);
+import {
+  getExerciseById,
+  generateStarRating,
+  isExerciseFavorite,
+  handleAddToRemoveToggle,
+} from './exerciseHelpers';
 
 export default async function renderExerciseById(exerciseId) {
   refs.modalEl.innerHTML = '';
@@ -30,7 +29,7 @@ export default async function renderExerciseById(exerciseId) {
     const starRatingMarkup = generateStarRating(rating);
 
     const modalMarkup = `<div class="modal-info">
-      <img class="exercise-img" src="${gifUrl}" alt="${name}" />
+      <img class="exercise-img" src="${gifUrl || 'Oops, there is no video for this exercise'}" alt="${name}" />
       <div class="modal-info-box">
         <h2 class="exercise-title">${name}</h2>
         <div class="exercise-rating">${rating.toFixed(
@@ -70,61 +69,5 @@ export default async function renderExerciseById(exerciseId) {
     }
   } catch (error) {
     console.error(error);
-  }
-}
-
-function generateStarRating(rating) {
-  const maxStars = 5;
-  let starsMarkup = '<ul class="star-rating">';
-  const roundedRating = Math.round(rating);
-
-  for (let i = 1; i <= maxStars; i++) {
-    if (roundedRating >= i) {
-      starsMarkup += `<li class="star full"><svg class="rate-star-icon" width="18" height="18"><use href="${spriteUrl}#icon-star"></use></svg></li>`;
-    } else {
-      starsMarkup += `<li class="star empty"><svg class="rate-star-icon" width="18" height="18"><use href="${spriteUrl}#icon-star"></use></svg></li>`;
-    }
-  }
-
-  starsMarkup += '</ul>';
-  return starsMarkup;
-}
-
-function isExerciseFavorite(exerciseId) {
-  const favorites = storageService.load('favorites') || [];
-  return favorites.some((exercise) => exercise._id === exerciseId);
-}
-
-function handleAddToRemoveToggle(event) {
-  const button = event.target;
-  const exerciseId = button.getAttribute('data-id');
-  const isFavorite = button.innerHTML.includes('Remove from favorites');
-
-  if (isFavorite) {
-    handleRemoveFromFavorites(exerciseId);
-  } else {
-    handleAddToFavorites(exerciseId);
-  }
-}
-
-function handleAddToFavorites(exerciseId) {
-  getExerciseById(exerciseId).then((exercise) => {
-    storageService.save('favorites', exercise);
-
-    const button = document.querySelector('.modal-btn');
-    if (button) {
-      button.innerHTML = `Remove from favorites
-        <svg width="18" height="18"><use href="${spriteUrl}#icon-trash"></use></svg>`;
-    }
-  });
-}
-
-function handleRemoveFromFavorites(exerciseId) {
-  storageService.remove('favorites', exerciseId);
-
-  const button = document.querySelector('.modal-btn');
-  if (button) {
-    button.innerHTML = `Add to favorites
-      <svg width="18" height="18"><use href="${spriteUrl}#icon-heart"></use></svg>`;
   }
 }
